requirements.txt
# Telegram Bot Framework
python-telegram-bot>=20.0

# Async HTTP Client
aiohttp>=3.9.0

# Environment Variables
python-dotenv>=1.0.0

# Payment Integrations (optional - uncomment when needed)
# stripe>=7.0.0
# alipay-sdk-python>=3.3.0

# Development Dependencies (optional)
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0
